actor: Skip top-levels in should_pick_paint()
When calling the clutter_actor_should_pick_paint() function, we should
also skip if the actor is a top-level.

The Stage is automatically cleared with the right color when picking,
as an optimization so that we don't have to submit further geometry or
change state; this means that calling should_pick_paint() should always
return FALSE on a Stage.