stage: Drop the custom pick() implementation
Now that clutter_actor_should_pick_paint() returns FALSE when called on
a top-level, we can drop the pick() override inside ClutterStage; the
default pick() implementation in ClutterActor calls should_pick_paint()
in order to decide whether or not the pick color should be used to
submit state and geometry. Since children of an Actor are painted
regardless of the pick color, the default pick() implementation inside
ClutterActor is now exactly the same as the override in ClutteStage.