actor: Add hit testing on ClutterActor
We should have a way to determine whether a point, defined in
actor-relative coordinates, is contained by an actor or not, as well as
determining the innermost child of a ClutterActor that contains the
point.

This requires some minimal matrix calculations to unproject a point
using the modelview and projection matrix tied to an actor, as well as a
new virtual function to allow actors to define non-rectangular shapes
for themselves.