Add future feature support
Python has an interesting concept: a __future__ module that defines
optional language features that are going to be mandatory, or dropped,
after a certain version is released. These features can be enabled by
the adventurous developers.

A similar mechanism for altering the default behaviour of Clutter would
complement nicely the deprecation mechanisms we have already in place.