actor: Introduce position, bounds, and frame
The current situation with the allocation, fixed position and size, and
margins inside ClutterActor is a bit messy, and the result of both an
inherited muddy implementation and a years of fudging around.

The allocation is defined to have a parent-relative origin, which takes
into account the parent-given position and the top left margins, and an
anti-origin with the parent-given size, minus the bottom right margins.
This means that users have to normalize the ClutterActorBox before using
it in layout and painting. Also, as a data structure, ClutterActorBox is
less than useful, as it doesn't lend itself easily to pixel-alignment.

Alongside with the allocation we also have a bunch of ancillary data
for the fixed positioning and fixed sizing that in a way compete with
the allocation, considering how our Do What I Mean™ API (methods and
properties) work; when asking for the position, size, or even preferred
size, we check for a series of flags, and return the appropriate value
depending on the state of the actor, leading to a convoluted (and hard
to follow and thus hard to maintain) code layout.

Finally, the API is at least asymmetric, with setters and getters and an
additional method called allocate() that does a ton of magic; and at
most baroque, with different, context-defined entry points and
behaviours for both actor writers and users.

Cleaning up this mess requires a series of steps, this commit being the
first one.

We introduce two separate storage units for the actor's position and
size:

  • Point position - the parent-relative origin of the actor;
  • Rect bounds    - the actor-relative paintable region of the actor;

We also add a "frame" ClutterRect, which is always computed on the fly
from the position, margins, and bounds of an actor.

When laying out the children of an actor, as well as when painting the
actor itself, we use "bounds", as it's the paintable 2D region of the
actor; the origin and size of the "bounds" rectangle take into account
the margins of an actor.

When allocating an actor, instead, we should set the "frame" rectangle;
the origin of the rectangle will determine the position, whereas the
size of the rectangle, along with the actor's margins, will be used to
determine the origin and size of the bounds of the actor.

Both bounds and frame have corresponding properties, which are also
animatable. This means that bounds and frame replace both the
clutter_actor_get_allocation_* family of functions, as well as the
clutter_actor_allocate() function, re-establishing symmetry.