actor: Move all state-related data into a separate structure
Instead of having all ancillary data structures littering the
ClutterActor source code, we can move them into a separate file.

This is the prelude to a more comprehensive encapsulation of the state
of a ClutterActor to an ancillary data structure, as well as the
separation between a "model" state and a "presentation" state.